Implement `Expression` and `Program`
While the parser parses S-Expressions only, the classes defined in
`program.h` are intented to turn these S-Expressions into expressions
of the F language. So far, only atoms are included. In the future,
there will be a class for function calls and for each special form.

`evaluate` currently does not take any arguments. In the future,
however, it will be extended with evaluation context.